fix(EditorWebView): Move local image upload to the editor web view
This makes sure that direct editing apps like text can also use regular
upload image methods through file input elements in the webview

fixes #11215
fixes #10744
